requirements.txt
# HackAI - Advanced Security Testing Framework
# Core dependencies
aiofiles
colorama

# Local LLM Support (Primary AI)
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
accelerate>=0.20.0

# Optional: Cloud AI (Fallback)
google-generativeai>=0.3.0
openai>=1.0.0

# Database
sqlite3  # Built-in with Python

# Network and security tools
requests>=2.31.0
urllib3>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web scraping and parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# JSON and configuration
pyyaml>=6.0
toml>=0.10.0

# CLI and interface
click>=8.1.0
rich>=13.0.0
prompt_toolkit>=3.0.0

# Utilities
python-dotenv>=1.0.0
psutil>=5.9.0
tqdm>=4.65.0

# Security and cryptography
cryptography>=41.0.0
paramiko>=3.3.0

# Optional: Advanced features
# pyserial>=3.5
# scapy>=2.5.0
